added some basic file io

+++ src/VM/interop/windows/window.cpp
@@ -171,7 +171,7 @@ namespace elsa {
 
 		void Window::render_text(int x, int y, const std::wstring& str)
 		{
-			TextOut(mem_hdc, x, y, str.c_str(), str.size());
+			TextOut(mem_hdc, x, y, str.c_str(), static_cast<int>(str.size()));
 		}
 	}
 }
